Create a single wlr_xcursor_manager
No need to create multiple of these: all use the same parameters.
