+++ server.h
@@ -70,6 +70,8 @@ struct cg_server {
 
 	struct wlr_foreign_toplevel_manager_v1 *foreign_toplevel_manager;
 
+	struct wlr_xcursor_manager *xcursor_manager;
+
 	bool xdg_decoration;
 	bool allow_vt_switch;
 	bool enable_xwayland;
